big reorganization and consolidation of views
